# 1.22x faster Greedy new lines