While adding new lines we can first check if ALL of them are valid. If so we can skip repeatedly calling ripper to check `CodeBlock#valid?`


Before: Finished in 1.43 seconds (files took 0.14836 seconds to load)
After:  Finished in 1.17 seconds (files took 0.15893 seconds to load)